+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [8.44] - 2026-04-17
+### Added
+- Added "Check for Updates" button to the Settings menu that fetches the latest release from GitHub to compare against current version.
+- Added version display `v{VERSION}` in Settings menu.
+
 ## [8.43] - 2026-04-17
 ### Changed
 - Swapped `setTimeout` debounce for `requestAnimationFrame` in the `FoodHelper` observer to ensure immediate, jitter-free UI updates when rendering food tables.
@@ -517,5 +522,3 @@ m- Included an interactive "View Stats" table modal showing lifetime consumption
 ## [7.41] - 2026-04-02
 ### Added
 - Living Area Mixer Link feature (added link to Mixer next to Hobo Grail/Kings Kiddie Cup/Golden Trolly).
-
-